Some UI fixes

+++ batch_export_unity.py
@@ -29,12 +29,12 @@ class UnityExporterPanel(Panel):
 
 	def draw(self, context):
 		layout = self.layout
-		mytool = context.scene.my_tools
+		myprops = context.scene.my_props
 
 		row = layout.box()
 		row.label(text="1)Select Unity Project Folder:")
 
-		row.prop(mytool, "project_path",icon ='FILE_IMAGE')
+		row.prop(myprops, "project_path",icon ='FILE_IMAGE')
 
 		row = layout.box()
 		row.label(text="2)Hit The Button To Export:")
@@ -42,7 +42,7 @@ class UnityExporterPanel(Panel):
 		row.operator("object.unity_batch_export")
 
 		row = layout.box()
-		row.prop(mytool,"sync_all", text="Sync Deleted Files")
+		row.prop(myprops,"sync_all", text="Sync Deleted Files")
 
 		row = layout.row()
 		row.label(text="Add * in Collection name to ignore.")
@@ -54,10 +54,13 @@ class UnityBatchExport(Operator):
 				def execute(self,context):
 					scene = context.scene
 					viewLayer = context.view_layer
-					collections = scene.collection.children
+					collectionLayer = bpy.context.view_layer.active_layer_collection
+					myprops = scene.my_props
+
+					bpy.ops.object.make_single_user(type='ALL', object=True, obdata=False, material=False, animation=False)
 #Check if project path is entered.
-					if scene.project_path.strip():
-						projectDir = bpy.path.abspath(scene.project_path)
+					if myprops.project_path.strip():
+						projectDir = bpy.path.abspath(myprops.project_path)
 					else:
 						raise Exception("No project path selected.")
 #Check if project path is valid
@@ -71,27 +74,29 @@ class UnityBatchExport(Operator):
 
 					bpy.ops.object.select_all(action='DESELECT')
 
-					for collection in collections:
-						collectionPath = exportdir +"/" + collection.name
+					for collection in collectionLayer.children:
+						if collection.is_visible:
+							collection = collection.collection
+							collectionPath = exportdir +"/" + collection.name
 
-						if len(collection.objects)==0 or "*" in collection.name:
-							collectionPath = collectionPath.replace("*","")
+							if len(collection.objects)==0 or "*" in collection.name:
+								collectionPath = collectionPath.replace("*","")
 
-							if os.path.isdir(collectionPath) and scene.sync_all:
-								shutil.rmtree(collectionPath)
-						else:
-							if not os.path.isdir(collectionPath):
-								os.mkdir(collectionPath)
+								if os.path.isdir(collectionPath) and scene.sync_all:
+									shutil.rmtree(collectionPath)
+							else:
+								if not os.path.isdir(collectionPath):
+									os.mkdir(collectionPath)
 
-							for obj in collection.objects:
-								obj.select_set(True)
-								name = obj.name
-								fn = os.path.join(collectionPath,name)
-								bpy.ops.object.transform_apply(location=False, rotation=True, scale=True)
-								bpy.context.object.hide_viewport = False
-								bpy.ops.export_scene.fbx(filepath=fn + ".fbx", use_selection=True, bake_space_transform=True,axis_forward="-Z",axis_up="Y",apply_scale_options="FBX_SCALE_ALL",check_existing=True,filter_glob="*.fbx",bake_anim=True,armature_nodetype='NULL',bake_anim_use_all_actions=True,embed_textures=False,object_types={'MESH'})
-								obj.select_set(False)
-								print("written:", fn)
+								for obj in collection.objects:
+									obj.select_set(True)
+									name = obj.name
+									fn = os.path.join(collectionPath,name)
+									bpy.ops.object.transform_apply(location=False, rotation=True, scale=True)
+									bpy.context.object.hide_viewport = False
+									bpy.ops.export_scene.fbx(filepath=fn + ".fbx", use_selection=True, bake_space_transform=True,axis_forward="-Z",axis_up="Y",apply_scale_options="FBX_SCALE_ALL",check_existing=True,filter_glob="*.fbx",bake_anim=True,armature_nodetype='NULL',bake_anim_use_all_actions=True,embed_textures=False,object_types={'MESH'})
+									obj.select_set(False)
+									print("written:", fn)
 
 					return{'FINISHED'}
 
@@ -100,12 +105,12 @@ classes = (MyProperties,UnityExporterPanel,UnityBatchExport)
 def register():
 	for c in classes:
 		bpy.utils.register_class(c)
-	bpy.types.Scene.my_tools = bpy.props.PointerProperty(type=MyProperties)
+	bpy.types.Scene.my_props = bpy.props.PointerProperty(type=MyProperties)
 
 def unregister():
 	for c in reversed(classes):
 		bpy.utils.unregister_class(c)
-	del bpy.types.Scene.my_tools
+	del bpy.types.Scene.my_props
 
 if __name__ == "__main__":
 	register()
